requirements.txt
asgiref==3.4.1
astroid==2.5.2
backports.entry-points-selectable==1.1.0
colorama==0.4.4
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers== 3.9.0
filelock==3.1.0
gunicorn==20.1.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
psycopg2==2.9.1
PyJWT==2.2.0
pylint==2.7.4
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.8.1
whitenoise==5.3.0
wrapt==1.12.1
